crawler/requirements.txt
beautifulsoup4
scapy
scrapy-splash
